devlog: update lod math

+++ LodGenerator/src/algorithms/basic_alg.cpp
@@ -11,14 +11,14 @@ namespace lod_generator{
 
         // !: Must be triangulated  
         // 2. Resize normals vector
-        std::vector<glm::vec3> normals;
+        std::vector<glm::dvec3> normals;
         normals.resize(cnt_faces);
 
         // Init Mesh Data
         mesh_data data = {};
         data.vertexes = std::shared_ptr<std::vector<double>>(&vertexes);
         data.indexes = std::shared_ptr<std::vector<uint32_t>>(&indexes);
-        data.normals = std::shared_ptr<std::vector<glm::vec3>>(&normals);
+        data.normals = std::shared_ptr<std::vector<glm::dvec3>>(&normals);
         
         data.valid_pairs = std::make_shared<std::vector<std::pair<uint32_t, uint32_t>>>();
         data.face_quadric_errors = std::make_shared<std::vector<glm::mat4x4>>();
